requirements.txt
fastapi
uvicorn
psycopg2-binary
gspread
pandas
python-dotenv
numpy
